syncutil: Move WaitGroupWithError from pkg util to pkg util/syncutil
